requirements.txt
chromadb
google-generativeai
flask
numpy
pandas
streamlit
